refactor: relocate deduplicator + category extractor
